#74 스프링 시큐리티를 적용하기

+++ src/main/java/com/saebyeok/saebyeok/service/ArticleService.java
@@ -2,7 +2,6 @@ package com.saebyeok.saebyeok.service;
 
 import com.saebyeok.saebyeok.domain.Article;
 import com.saebyeok.saebyeok.domain.ArticleRepository;
-import com.saebyeok.saebyeok.domain.Gender;
 import com.saebyeok.saebyeok.domain.Member;
 import com.saebyeok.saebyeok.dto.ArticleCreateRequest;
 import com.saebyeok.saebyeok.dto.ArticleResponse;
@@ -17,7 +16,6 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDateTime;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -26,12 +24,13 @@ import java.util.stream.Collectors;
 public class ArticleService {
 
     public static final int LIMIT_DAYS = 7;
+    private static final String NOT_YOUR_ARTICLE_MESSAGE = "자신의 게시글이 아닙니다!";
 
     private final ArticleRepository articleRepository;
     private final ArticleEmotionService articleEmotionService;
     private final ArticleSubEmotionService articleSubEmotionService;
 
-    public List<ArticleResponse> getArticles(int page, int size) {
+    public List<ArticleResponse> getArticles(Member member, int page, int size) {
         Pageable pageable = PageRequest.of(page, size, Sort.by(Sort.Direction.DESC, "id"));
 
         return articleRepository.findAllByCreatedDateGreaterThanEqual(LocalDateTime.now().minusDays(LIMIT_DAYS), pageable).
@@ -39,17 +38,15 @@ public class ArticleService {
                 map(article -> {
                     EmotionResponse emotionResponse = articleEmotionService.findEmotion(article);
                     List<SubEmotionResponse> subEmotionResponses = articleSubEmotionService.findSubEmotions(article);
-                    return new ArticleResponse(article, emotionResponse, subEmotionResponses);
+                    return new ArticleResponse(article, member, emotionResponse, subEmotionResponses);
                 }).
                 collect(Collectors.toList());
     }
 
     @Transactional
     public Article createArticle(Member member, ArticleCreateRequest request) {
-        // Todo: 초기 개발을 위한 member 생성 로직 삭제하고 security 적용하면 member 받아오기
-        Member test = new Member(1L, 1991, Gender.FEMALE, LocalDateTime.now(), false, new ArrayList<>());
         Article article = request.toArticle();
-        article.setMember(test);
+        article.setMember(member);
 
         articleEmotionService.createArticleEmotion(article, request.getEmotionId());
         articleSubEmotionService.createArticleSubEmotion(article, request.getSubEmotionIds());
@@ -57,22 +54,24 @@ public class ArticleService {
         return articleRepository.save(article);
     }
 
-    public ArticleResponse readArticle(Long articleId) {
+    public ArticleResponse readArticle(Member member, Long articleId) {
         Article article = articleRepository.findByIdAndCreatedDateGreaterThanEqual(articleId, LocalDateTime.now().minusDays(LIMIT_DAYS))
                 .orElseThrow(() -> new ArticleNotFoundException(articleId));
         EmotionResponse emotionResponse = articleEmotionService.findEmotion(article);
         List<SubEmotionResponse> subEmotionResponses = articleSubEmotionService.findSubEmotions(article);
 
-        return new ArticleResponse(article, emotionResponse, subEmotionResponses);
+        return new ArticleResponse(article, member, emotionResponse, subEmotionResponses);
     }
 
     @Transactional
-    public void deleteArticle(Long id) {
-        Article article = articleRepository.findById(id)
-                .orElseThrow(() -> new ArticleNotFoundException(id));
-
+    public void deleteArticle(Member member, Long articleId) throws IllegalAccessException {
+        Article article = articleRepository.findById(articleId)
+                .orElseThrow(() -> new ArticleNotFoundException(articleId));
+        if (!article.isWrittenBy(member)) {
+            throw new IllegalAccessException(NOT_YOUR_ARTICLE_MESSAGE);
+        }
         articleEmotionService.deleteArticleEmotion(article);
         articleSubEmotionService.deleteArticleSubEmotion(article);
-        articleRepository.deleteById(article.getId());
+        articleRepository.deleteById(articleId);
     }
 }
